# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: departmentIds
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated department ids; matches employees in any of them
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EmployeeStatus"
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Employee sections ────────────────────────────────────────────────────
...
components:
  schemas:
...
    EmploymentType:
      type: string
      description: Employment type value — managed via List of Values
    EmployeeStatus:
      type: string
      enum: [active, inactive, on_leave, leaver]
...
    Employee:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - jobTitle
        - departments
        - employmentType
        - status
        - startDate
        - salary
        - avatarUrl
        - leaverReason
        - leaverDate
        - createdAt
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phones:
          type: array
          items:
            $ref: "#/components/schemas/PhoneSummary"
          description: Phone numbers for this employee (only present on single-record responses)
        jobTitle:
          type: string
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentSummary"
        employmentType:
          $ref: "#/components/schemas/EmploymentType"
        status:
          $ref: "#/components/schemas/EmployeeStatus"
        startDate:
          type: string
          format: date
        salary:
          type: ["number", "null"]
        avatarUrl:
          type: ["string", "null"]
        leaverReason:
          type: ["string", "null"]
          description: LOV value from leaver_reason category; required when status is leaver
        leaverDate:
          type: ["string", "null"]
          format: date
          description: Date the employee left; defaults to today when status is set to leaver
        createdAt:
          type: string
          format: date-time
        pendingDisclosureReview:
          type: boolean
          description: >
            True when the employee has at least one disclosure with conviction details
            that has not yet been signed off. Only populated for users with
            view_disclosures or sysadmin permission; false otherwise.
    EmployeeInput:
      type: object
      required: [firstName, lastName, email, jobTitle, employmentType, status, startDate, userRoleIds, temporaryPassword]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        jobTitle:
          type: string
          minLength: 1
        departmentIds:
          type: array
          items:
            type: integer
          default: []
        employmentType:
          $ref: "#/components/schemas/EmploymentType"
        status:
          $ref: "#/components/schemas/EmployeeStatus"
        startDate:
          type: string
          format: date
        salary:
          type: ["number", "null"]
        avatarUrl:
          type: ["string", "null"]
        userRoleIds:
          type: array
          items:
            type: integer
          minItems: 1
          description: Role IDs to assign to the automatically created user account
        temporaryPassword:
          type: string
          minLength: 8
          description: Temporary password for the automatically created user account
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        jobTitle:
          type: string
          minLength: 1
        departmentIds:
          type: array
          items:
            type: integer
        employmentType:
          $ref: "#/components/schemas/EmploymentType"
        status:
          $ref: "#/components/schemas/EmployeeStatus"
        startDate:
          type: string
          format: date
        salary:
          type: ["number", "null"]
        avatarUrl:
          type: ["string", "null"]
        leaverReason:
          type: ["string", "null"]
          description: Required when setting status to leaver
        leaverDate:
          type: ["string", "null"]
          format: date
          description: Defaults to today if omitted when setting status to leaver
...
    DepartmentSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    PhoneLabel:
      type: string
      enum: [Mobile, Home, Work, Other]
    PhoneSummary:
      type: object
      required: [id, number, label, isPrimary]
      properties:
        id:
          type: integer
        number:
          type: string
        label:
          $ref: "#/components/schemas/PhoneLabel"
        isPrimary:
          type: boolean
...