# lib/db/src/schema/hr/employees.ts
import {
  pgTable,
  serial,
  text,
  integer,
  numeric,
  date,
  timestamp,
} from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const employmentTypeValues = [
  "full_time",
  "part_time",
  "contract",
  "intern",
] as const;

export const employeeStatusValues = ["active", "inactive", "on_leave", "leaver"] as const;

export const employeesTable = pgTable("employees", {
  id: serial("id").primaryKey(),
  firstName: text("first_name").notNull(),
  lastName: text("last_name").notNull(),
  email: text("email").notNull(),
  jobTitle: text("job_title").notNull(),
  employmentType: text("employment_type").notNull(),
  status: text("status", { enum: employeeStatusValues })
    .notNull()
    .default("active"),
  startDate: date("start_date", { mode: "string" }).notNull(),
  salary: numeric("salary", { precision: 12, scale: 2 }),
  avatarUrl: text("avatar_url"),
  leaverReason: text("leaver_reason"),
  leaverDate: date("leaver_date", { mode: "string" }),
  createdAt: timestamp("created_at", { withTimezone: true })
    .notNull()
    .defaultNow(),
});

export const insertEmployeeSchema = createInsertSchema(employeesTable).omit({
  id: true,
  createdAt: true,
});
export type InsertEmployee = z.infer<typeof insertEmployeeSchema>;
export type Employee = typeof employeesTable.$inferSelect;